server/requirements.txt
amqp>=2.3.2,<2.4
appdirs>=1.4.3,<1.5
atomicwrites>=1.1.5,<1.3
attrs>=18.1.0,<19
backports.functools-lru-cache==1.5; python_version == '2.7'
billiard>=3.5.0.3,<3.6
boto>=2.48.0,<2.50
boto3==1.7.78
celery>=4.1.1,<4.3
certifi>=2018.4.16
chardet>=3.0.4,<3.1
configparser>=3.5.0,<3.6; python_version == '2.7'
coverage==4.5.2
Django>=1.11.16,<1.12
django-chartjs==1.3
djangorestframework>=3.8.2,<3.9
enum34>=1.1.6,<1.2; python_version == '2.7'
fasteners>=0.14.1,<0.15
flake8>=3.5.0,<3.7
funcsigs>=1.0.2,<1.1; python_version == '2.7'
futures>=3.2.0,<3.3; python_version == '2.7'
idna>=2.6,<2.8
isort>=4.3.4,<4.4
kombu>=4.2.1,<4.3
laniakea==1.15.0
lazy-object-proxy>=1.3.1,<1.4
mccabe>=0.6.1,<0.7
mock==2.0.0; python_version == '2.7'
monotonic==1.5
more-itertools>=4.2.0,<4.4
numpy>=1.14.3,<2
pluggy>=0.7.0,<0.8
py>=1.5.3
pycodestyle>=2.4,<2.5
pyflakes>=2.0.0,<2.1
pytest>=3.10.0,<4.0
pytest-cov>=2.6.0,<2.7
pytest-django>=3.4.4,<3.5
pytest-flake8>=1.0.1,<1.1
pytest-pythonpath>=0.7.2,<0.8
pytz>=2018.4
redis>=2.10.6,<2.11
requests==2.20.1
singledispatch>=3.4.0.3,<3.4.1; python_version == '2.7'
six>=1.12.0,<1.13
urllib3>=1.22,<2
vine>=1.1.4,<1.2
wrapt>=1.10.11,<1.11
#django-livesync==0.4
